==> analysis/Organoid_scRNA_analysis/03-DEG/01-Cluster_DEG_2025_05_29.Rmd <==
---
title: "Cluster_DEG"
author: "Ximing Ran"
date: "2025-05-29"
output:
  pdf_document:
    toc: true
    toc_depth: 3
    number_sections: false
    latex_engine: xelatex
  html_document:
    # code_folding: hide
    toc: true
    toc_float: true
hitheme: tomorrow
highlighter: highlight.js

---

```{r setup, include=FALSE}
# load libraries
library(tidyverse)
library(knitr)
set.seed(2025)

knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE,
  message = FALSE,
  fig.width = 9,
  fig.height = 4,
  fig.path = "result/01-Cluster_DEG/Analysis_plot"
)

theme_set(theme_bw() + theme(legend.position = "bottom"))
knitr::kable(head(mtcars[, 1:4]), "simple")
```

```{r}
library(tibble)
library(tidyr)
library(dplyr)
library(rtracklayer)
library(dplyr)
# library(Seurat)
library(Signac)
library(EnsDb.Hsapiens.v86)
library(ggplot2)
library(cowplot)
library(simspec)
library(cowplot)
library(AnnotationHub)
library(Seurat)
library(tidyverse)
library(openxlsx)
library(dplyr)
library(conflicted)
library(harmony)
library(data.table)
library(pheatmap)
library(ggplot2)
library(patchwork)

library(knitr)
library(kableExtra)

conflict_prefer("filter", "dplyr")  # Prefer dplyr's filter()
conflict_prefer("lag", "dplyr")     # Prefer dplyr's lag()

```



```{r local_function_load}
# load function from local files
# source(here::here("source", "sc_functions.R"))
```



\newpage

# 1. Read the merged Data
```{r load_merged_data}
merged <- readRDS(here::here("data", "synaptosomes_scRNA", "04_merged_with_clustering_4_mannual_correct.rds"))
```


```{r}

####################### cluster DEG
merged$orig.group <- ifelse(merged$orig.ident %in% c("C1", "C2"), "CTRL", "22q")
merged$deg_group <- paste0(merged$orig.group, "_", merged$seurat_clusters_merge)
table(merged$deg_group)

DefaultAssay(merged) <- "RNA"
Idents(merged) <- "deg_group"
lapply(unique(merged$seurat_clusters_merge), function(cl){
  print(cl)
  g1 <- paste0("CTRL_", cl)
  g2 <- paste0("22q_", cl)
  deg <- FindMarkers(merged, ident.1 = g1, ident.2 = g2)
  deg$cluster <- cl
  deg$gene <- rownames(deg)
  deg
}) -> deg_list
names(deg_list) <- paste0(unique(merged$seurat_clusters_merge))

# make. avg_log2FC the -avg_log2FC
deg_list <- lapply(deg_list, function(deg){
  deg$avg_log2FC <- -deg$avg_log2FC
  deg
})




library(stringr)
# assume deg_list is a named list, one element per cluster
old_names <- names(deg_list)
# truncate every name to at most 31 characters (no “…”)
new_names <- str_trunc(old_names, width = 31, side = "right", ellipsis = "")
names(deg_list) <- new_names

# now write
openxlsx::write.xlsx(
  deg_list,
  "result/01-Cluster_DEG/cluster_DEGs.xlsx"
)


```


```{r DEG_summary, fig.width=8, fig.height=4}

library(dplyr)
library(tidyr)
library(ggplot2)
library(purrr)

# your vector of logFC cutoffs
logFCs <- c(0.1, 0.25, 0.5, 1)

for (th in logFCs) {
  # 1. tally up/down per cluster at this threshold
  counts_df <- imap_dfr(deg_list, ~{
    df2 <- .x %>%
      filter(p_val_adj < 0.05, abs(avg_log2FC) > th)
    tibble(
      cluster = .y,
      up      = sum(df2$avg_log2FC >  0, na.rm = TRUE),
      down    = sum(df2$avg_log2FC <  0, na.rm = TRUE)
    )
  })
  
  # 2. pivot longer and make 'down' negative
  counts_long <- counts_df %>%
    pivot_longer(c(up, down), names_to = "direction", values_to = "n") %>%
    mutate(
      n         = ifelse(direction == "down", -n, n),
      direction = factor(direction, levels = c("down","up"))
    )
  
  # figure out padding
  pad <- max(abs(counts_long$n)) * 0.1
  
  # 3. build the plot
  p <- ggplot(counts_long, aes(x = n, y = cluster, fill = direction)) +
    geom_col(width = 0.7) +
    geom_text(
      aes(
        x     = n,
        label = abs(n),
        hjust = ifelse(direction == "up", -0.1, 1.1)
      ),
      size = 3
    ) +
    scale_x_continuous(
      labels = abs,
      # add extra room on both sides (you may need to tweak these multipliers)
      expand = expansion(add = c(max(abs(counts_long$n)) * 0.2,
                                 max(abs(counts_long$n)) * 0.2))
    ) +
    scale_fill_manual(
      values = c(down = "steelblue", up = "firebrick"),
      labels = c("Down-regulated","Up-regulated")
    ) +
    labs(
      title = paste0("DEG counts (|logFC| > ", th, ")"),
      x     = "Number of DE genes",
      y     = "Cell type (cluster)",
      fill  = "Direction"
    ) +
    theme_classic(base_size = 14) +
    theme(
      legend.position     = "top",
      axis.text.y         = element_text(size = 10),
      # bump out the right margin (t, r, b, l)
      plot.margin         = margin(5, 30, 5, 5, "pt")
    ) +
    # allow drawing of labels beyond the panel
    coord_cartesian(clip = "off")
  print(p)
  
  # 4. save it
  ggsave(
    filename = paste0("result/01-Cluster_DEG/diverging_bar_logFC_", th, ".png"),
    plot     = p,
    width    = 8, 
    height   = 6,
    dpi      = 300
  )
}






```


```{r}




# library(dplyr)
# library(purrr)
# library(gprofiler2)
# 
# # your vector of log₂FC thresholds:
# logFCs <- c(1, 0.5, 0.25)
# 
# for (th in logFCs) {
#   message("Running enrichment for |log2FC| > ", th)
#   # make a clean directory for this threshold
#   outdir <- file.path("result/01-Cluster_DEG",
#                       paste0("cluster_DEG_enrichment_logFC_", th))
#   dir.create(outdir, recursive = TRUE, showWarnings = FALSE)
#   
#   # filter each cluster once
#   deg_list_sign <- map(deg_list, ~
#                          .x %>%
#                          filter(p_val_adj < 0.05, abs(avg_log2FC) > th)
#   )
#   
#   # run enrichment per cluster
#   walk(deg_list_sign, function(deg) {
#     # get a safe name
#     cluster_name <- unique(deg$cluster)
#     safe_name    <- gsub("\\s+", "_", cluster_name)
#     
#     # gene sets
#     up_genes   <- rownames(filter(deg, avg_log2FC >  0))
#     down_genes <- rownames(filter(deg, avg_log2FC <  0))
#     
#     # run gost
#     up_enr   <- gost(up_genes,   organism = "hsapiens",
#                      correction_method = "fdr", evcodes = TRUE)
#     
#     # write results
#     write.csv(
#       dplyr::select(up_enr$result, -parents),
#       file = file.path(outdir,
#                        paste0(safe_name, "_up_enrichment.csv")),
#       row.names = FALSE
#     )
#     
#     
#     down_enr <- gost(down_genes, organism = "hsapiens",
#                      correction_method = "fdr", evcodes = TRUE)
#     
#     write.csv(
#       dplyr::select(down_enr$result, -parents),
#       file = file.path(outdir,
#                        paste0(safe_name, "_down_enrichment.csv")),
#       row.names = FALSE
#     )
#   })
#   message("Done enrichment at |log2FC| > ", th)
# }
```


\newpage
Session Information
```{r session_info}
sessionInfo()
```
