==> analysis/Organoid_scRNA_analysis/03-DEG/03-Figure_2025_06_03.Rmd <==
---
title: "Figrue for 2025-06-03"
author: "Ximing Ran"
date: "2025-06-03"
output:
  pdf_document:
    toc: true
    toc_depth: 3
    number_sections: false
    latex_engine: xelatex
  html_document:
    # code_folding: hide
    toc: true
    toc_float: true
hitheme: tomorrow
highlighter: highlight.js

---

```{r setup, include=FALSE}
# load libraries
library(tidyverse)
library(knitr)
set.seed(2025)

knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE,
  message = FALSE,
  fig.width = 9,
  fig.height = 4,
  fig.path = "result/03-Figure-06-03/Analysis_plot/"
)

theme_set(theme_bw() + theme(legend.position = "bottom"))
knitr::kable(head(mtcars[, 1:4]), "simple")
```

```{r}
library(tibble)
library(tidyr)
library(dplyr)
library(rtracklayer)
library(dplyr)
# library(Seurat)
library(Signac)
library(EnsDb.Hsapiens.v86)
library(ggplot2)
library(cowplot)
library(simspec)
library(cowplot)
library(AnnotationHub)
library(Seurat)
library(tidyverse)
library(openxlsx)
library(dplyr)
library(conflicted)
library(harmony)
library(data.table)
library(pheatmap)
library(ggplot2)
library(patchwork)
library(extrafont)

library(knitr)
library(kableExtra)
library(Seurat)
library(ggplot2)
library(dplyr)

conflict_prefer("filter", "dplyr")  # Prefer dplyr's filter()
conflict_prefer("lag", "dplyr")     # Prefer dplyr's lag()

```

```{r local_function_load}
# load function from local files
# source(here::here("source", "sc_functions.R"))
```


\newpage
# 1. Read the merged Data
```{r load_merged_data}
seruat_obj <- readRDS(here::here("data", "synaptosomes_scRNA", "04_merged_with_clustering_4_mannual_correct.rds"))


```


# 2. Plot the UMAP
```{r UMAP_plot , fig.width=8, fig.height=5}
Idents(seruat_obj) <- seruat_obj$seurat_clusters_merge
# table(seruat_obj$seurat_clusters_merge)

# swap the      Upper layer excitatory neurons  and Excitatory neurons 
seruat_obj$seurat_clusters_merge <- recode(
  seruat_obj$seurat_clusters_merge,
  "Upper layer excitatory neurons" = "Excitatory neurons",
  "Excitatory neurons" = "Upper layer excitatory neurons"
)


seruat_obj$cluster_num <-as.integer(Idents(seruat_obj))-1

# 1. Grab your existing cell-type names
orig_types <- levels(Idents(seruat_obj))

# 2. Build a mapping from cell type → number
num_map <- setNames(seq_along(orig_types)-1, orig_types)

seruat_obj$cluster_annot <- factor(
  paste0(seruat_obj$cluster_num, "-", as.character(seruat_obj$seurat_clusters_merge))
)

table(seruat_obj$cluster_annot)

# recode the 22Q into the 22q11DS

seruat_obj$group_annot <- recode(
  seruat_obj$group,
  "22Q" = "22q11DS"
) 
# relevel the factor to have CTRL first
seruat_obj$group_annot <- factor(
  seruat_obj$group_annot,
  levels = c("CTRL", "22q11DS")
)





Idents(seruat_obj) <- "cluster_annot"

# 5. Base UMAP (no labels) with transparent backgrounds
umap_plot <- DimPlot(
  seruat_obj,
  reduction = "umap",
  # group.by = "cluster_annot", 
  # split.by = "group_annot",
  label   = FALSE,
  repel   = TRUE
) 

# subset the part before the -
umap_plot$data$num <- sub("-.*", "", umap_plot$data$ident)
umap_plot$data$num <- factor(umap_plot$data$num, levels = seq(0, 9))

# 3) Call LabelClusters with id = "ident"
umap_plot_labeled <- LabelClusters(
  plot     = umap_plot,
  id       = "num",   # the data column that holds "0".."9"
  # clusters = src,       # which groups to annotate
  # labels   = anno,      # what text to put in place of each "0".."9"
  repel    = FALSE,
  size     = 4
) 

umap_plot_labeled <- umap_plot_labeled +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA)
  )


ggsave(
  filename = "result/03-Figure-06-03/UMAP_labeled_annotated_all.png",
  plot     = umap_plot_labeled,
  width    = 8,
  height   = 5,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)

# save to pdf
ggsave(
  filename = "result/03-Figure-06-03/UMAP_labeled_annotated_all.pdf",
  plot     = umap_plot_labeled,
  width    = 8,
  height   = 5
)

print(umap_plot_labeled)
```

```{r UMAP_plot_group , fig.width=12, fig.height=5}
# 4. Use the new annotated factor as your Idents()
Idents(seruat_obj) <- "cluster_annot"

# 5. Base UMAP (no labels) with transparent backgrounds
umap_plot <- DimPlot(
  seruat_obj,
  reduction = "umap",
  # group.by = "cluster_annot", 
  split.by = "group_annot",
  label   = FALSE,
  repel   = TRUE
) 

# subset the part before the -
umap_plot$data$num <- sub("-.*", "", umap_plot$data$ident)
umap_plot$data$num <- factor(umap_plot$data$num, levels = seq(0, 9))

# 3) Call LabelClusters with id = "ident"
umap_plot_labeled <- LabelClusters(
  plot     = umap_plot,
  id       = "num",   # the data column that holds "0".."9"
  # clusters = src,       # which groups to annotate
  # labels   = anno,      # what text to put in place of each "0".."9"
  repel    = FALSE,
  size     = 4
) 

umap_plot_labeled <- umap_plot_labeled +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA)
  )

ggsave(
  filename = "result/03-Figure-06-03/UMAP_labeled_annotated.png",
  plot     = umap_plot_labeled,
  width    = 12,
  height   = 5,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)

# save to pdf
ggsave(
  filename = "result/003-Figure-06-03/UMAP_labeled_annotated.pdf",
  plot     = umap_plot_labeled,
  width    = 12,
  height   = 5
)



print(umap_plot_labeled)
```

# 3. Plot the cell_type proportion
```{r cell_type_proportion, fig.width=8, fig.height=5}
library(dplyr)


seruat_obj@meta.data %>%
  group_by(cluster_num,group_annot) %>%
  count() %>%
  group_by(group_annot) %>%
  mutate(percent = n / sum(n) * 100) %>%
  mutate(group = factor(group_annot, levels = c("CTRL", "22q11DS")),
         x = as.numeric(cluster_num)) %>%
  ggplot() +
  geom_bar(aes(x = x, y = y),
           stat = "identity", fill = "pink", alpha = 0.5,
           width = 1,
           data = data.frame(x = c(0, 3, 4, 7, 9),
                             y = 30)) +
  geom_bar(aes(x = x, y = percent,
               group = group,fill = group),
           width = 0.8,
           stat = "identity", position = position_dodge()) +
  scale_fill_manual(values = c("#00BFC4", "#F8766D")) +
  scale_y_continuous(limits = c(-4, 30), expand = expansion(add = c(0, 0))) +
  geom_point(mapping = aes(x, y, color = as.factor(x)),
             size = 7,
             data = data.frame(x = 0:9, y = -2),
             show.legend = F) +
  geom_text(mapping = aes(x, y, label = x),
            data = data.frame(x = 0:9, y = -2),
            show.legend = F) +
  geom_hline(yintercept = 0, linewidth = 0.2) +
  xlab("Cluster") + ylab("Percentage") +
  theme_bw(base_size = 12, base_family = "Arial") +
  # theme_classic(base_size = 12, base_family = "Arial") +
  theme(axis.text = element_text(color = "black"),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank()
  ) +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA)
  ) -> p

print(p)


ggsave(
  filename = "result/03-Figure-06-03/cell_type_proportion.png",
  plot     = p,
  width    = 8,
  height   = 5,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)

# save to pdf
ggsave(
  filename = "result/03-Figure-06-03/cell_type_proportion.pdf",
  plot     = p,
  width    = 8,
  height   = 5,
  bg       = "transparent" 
)



```


# 4. Plot the marker
```{r marker_plot, fig.width=8, fig.height=6}
marker_df <- read.csv(here::here("data", "synaptosomes_scRNA", "ct_mannual_0603.csv"),
                       stringsAsFactors = FALSE)

# Split on commas, then unlist to get a simple vector of genes
all_genes <- unlist(strsplit(as.character(marker_df$marker2), split = ","))

# (Optional) Trim any leading/trailing whitespace
all_genes <- trimws(all_genes)
all_genes <- unique(all_genes)


# Plot the DotPlot
Idents(seruat_obj) <- "cluster_annot"
DefaultAssay(seruat_obj) <- "RNA"
DotPlot(
  seruat_obj,
  features = all_genes,
  group.by = "cluster_num",
  # split.by = "group_annot",
  # cols = c("lightgrey", "blue"),
  # scale.min = 0.1,
  # scale.max = 0.8
) +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA)
  ) + coord_flip() ->p
ggsave(
  filename = "result/03-Figure-06-03/marker_plot.png",
  plot     = p,
  width    = 8,
  height   = 6,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)
ggsave(
  filename = "result/03-Figure-06-03/marker_plot.pdf",
  plot     = p,
  width    = 8,
  height   = 6,
  bg       = "transparent" 
)



seruat_obj$cluster_annot <- recode(
  seruat_obj$cluster_annot, "9-Immature preoptic area (pOA) interneurons"= "9-Immature pOA interneurons")

table(seruat_obj$cluster_annot)

DotPlot(
  seruat_obj,
  features = all_genes,
  group.by = "cluster_annot",
  # split.by = "group_annot",
  # cols = c("lightgrey", "blue"),
  # scale.min = 0.1,
  # scale.max = 0.8
) + coord_flip() +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA),
    axis.text.x       = element_text(angle = 45, hjust = 1)
  ) ->p 
ggsave(
  filename = "result/03-Figure-06-03/marker_plot_annot.png",
  plot     = p,
  width    = 8,
  height   = 8,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)
ggsave(
  filename = "result/03-Figure-06-03/marker_plot_annot.pdf",
  plot     = p,
  width    = 8,
  height   = 8,
  bg       = "transparent" 
)









```

```{r marker_plot_vln, fig.width=6, fig.height=8}
Idents(seruat_obj) <- factor(
  seruat_obj$cluster_num,
  levels = c("0","1","2","3","4","5","6","7","8","9")
)
all_genes_rev <- rev(all_genes)
VlnPlot(object = seruat_obj, features =all_genes_rev , stack = TRUE, sort = FALSE, flip = TRUE) +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA),
    legend.position = "none") ->p

ggsave(
  filename = "result/03-Figure-06-03/marker_plot_vln.png",
  plot     = p,
  width    = 6,
  height   = 6,
  dpi      = 300,
  bg       = "transparent"   # <-- this makes the PNG background transparent
)

ggsave(
  filename = "result/03-Figure-06-03/marker_plot_vln.pdf",
  plot     = p,
  width    = 6,
  height   = 6,
  bg       = "transparent" 
)
Idents(seruat_obj) <- "cluster_annot"
VlnPlot(object = seruat_obj, features = all_genes_rev , stack = TRUE, sort = FALSE, flip = TRUE) +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA),
    plot.margin       = unit(c(0, 0, 0, 0.8), "in"),
    legend.position = "none") ->p

ggsave(
  filename = "result/03-Figure-06-03/marker_plot_vln_annot.png",
  plot     = p,
  width    = 6,
  height   = 8,
  dpi      = 300,
  bg       = "transparent"   
)
ggsave(
  filename = "result/03-Figure-06-03/marker_plot_vln_annot.pdf",
  plot     = p,
  width    = 6,
  height   = 8,
  bg       = "transparent" 
)


```


\newpage
# 5. Plot for deleted region
```{r 22q_plot, fig.width=8, fig.height=10}
target_22q_gene <- read.csv(here::here("data","ref" ,"22q_gene_2024_10_17.csv"))
target_gene <- target_22q_gene$gene
target_gene <- target_gene[1:(length(target_gene) - 4)]
target_gene <- rev(target_gene)

# count the gene expression by 
seruat_obj_22q <- subset(seruat_obj, features = target_gene)
unique(seruat_obj_22q$orig.ident)

# summary by orid.

DotPlot(
  seruat_obj_22q,
  features = target_gene ,
  group.by = "orig.ident",
  # split.by = "group_annot",
  # cols = c("lightgrey", "blue"),
  # scale.min = 0.1,
  # scale.max = 1
) + coord_flip() +
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA),
    axis.text.x       = element_text(angle = 45, hjust = 1)
  )



VlnPlot(
  seruat_obj_22q,
  features = target_gene ,
  group.by = "orig.ident",
  stack = TRUE, 
  sort = FALSE,
  flip = TRUE
  # split.by = "group_annot",
  # cols = c("lightgrey", "blue"),
  # scale.min = 0.1,
  # scale.max = 0.8
) +  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA),
    plot.margin       = unit(c(0, 0, 0, 0.8), "in"),
    legend.position = "none") 




```

```{r}
plot_gene_heatmap <- function(counts_mat,
                              annotation_col,
                              gene_list,
                              figure_folder,
                              file_name,
                              reference_group,
                              compare_group,
                              cluster_rows   = FALSE,
                              cluster_cols   = FALSE,
                              save           = TRUE,
                              fig.width      = 8,
                              fig.height     = 8,
                              scale          = "none",
                              show_rownames  = FALSE) {
  #
  # counts_mat       : a numeric matrix of raw counts (rows = genes, columns = samples)
  # annotation_col   : a data.frame with one column named “group” (rownames = sample names)
  # gene_list        : character vector of gene IDs (must match rownames(counts_mat))
  # figure_folder    : where to save PNG/PDF
  # file_name        : base name (without extension)
  # reference_group, compare_group : two levels used in annotation_col$group
  # cluster_rows/cluster_cols     : logical for pheatmap clustering
  # scale            : "none" / "row" / "column"
  # show_rownames    : logical
  #
  
  # 1. Build a blue-red palette (RdBu reversed):
  brewer_palette <- "RdBu"
  ramp <- colorRampPalette(RColorBrewer::brewer.pal(11, brewer_palette))
  mr <- ramp(256)[256:1]
  
  # 2. Restrict counts_mat to gene_list and to samples in annotation_col
  all_samples <- rownames(annotation_col)
  if (!all(all_samples %in% colnames(counts_mat))) {
    stop("Some rownames(annotation_col) are not column names of counts_mat")
  }
  subset_mat <- counts_mat[gene_list, all_samples, drop = FALSE]
  
  # 3. Center each gene (row) to zero mean if scale = "none"
  #    If scale = "row", let pheatmap handle it.
  if (scale == "none") {
    row_means <- rowMeans(subset_mat, na.rm = TRUE)
    subset_mat <- subset_mat - row_means
  }
  
  # 4. Build annotation colors for the two groups
  unique_groups <- c(reference_group, compare_group)
  color_palette <- c("#10d5da", "#fe867f")
  assigned_colors <- setNames(
    color_palette[seq_along(unique_groups)],
    unique_groups
  )
  annotation_colors <- list(group = assigned_colors)
  
  # 5. Ensure annotation_col has exactly one column named “group”
  if (!"group" %in% colnames(annotation_col)) {
    stop("annotation_col must have a column named 'group'")
  }
  annotation_df <- annotation_col[, "group", drop = FALSE]
  
  # 6. Draw the heatmap
  p <- pheatmap::pheatmap(
    mat               = subset_mat,
    color             = mr,
    scale             = scale,
    annotation_col    = annotation_df,
    annotation_colors = annotation_colors,
    cluster_rows      = cluster_rows,
    cluster_cols      = cluster_cols,
    fontsize_col      = 10,
    fontsize_row      = 10,
    treeheight_row    = if (cluster_rows) 50 else 0,
    show_rownames     = show_rownames,
    border_color      = NA,
    angle_col         = 0
  )
  
  
  # 7. Save to PNG/PDF if requested
  if (save) {
    png_path <- file.path(figure_folder, paste0(file_name, ".png"))
    pdf_path <- file.path(figure_folder, paste0(file_name, ".pdf"))
    ggsave(png_path,  plot=p, width = fig.width, height = fig.height, units = "in", dpi = 300)
    ggsave(pdf_path,plot=p,  width = fig.width, height = fig.height, units = "in")
  }
  
  return(p)
}

```



```{r 22q_scRNA_heatmap, fig.width=6, fig.height=10}
target_cluster <- c(0,1,5,8,9)
Idents(seruat_obj_22q) <- "cluster_num"

annotation_col <- data.frame(
  group = c("CTRL", "CTRL", "22q11DS", "22q11DS"),
  row.names = c("006", "008", "22q002", "22q005")
)


for (target in target_cluster){
  print(paste0("Cluster ", target))
  # turn target into character
  # target <- as.character(target)
  # seruat_sub <- subset(x=  seruat_obj_22q, subset = cluster_num ==0)

  # (2) Run AverageExpression on the “counts” slot, grouping by orig.ident:
  # avg_counts_list <- AverageExpression(
  #   seruat_sub,
  #   assays   = "RNA",
  #   slot     = "counts",
  #   group.by = "orig.ident"
  # )
  # 
  # # (3) Extract the resulting matrix (genes × orig.ident):
  # avg_counts_mat <- avg_counts_list$RNA
  
  # 1a) Identify all cell barcodes in the target cluster:
cell_id <- WhichCells(seruat_obj_22q, expression = cluster_num == target)

# 1b) Create a data.frame with each cell’s barcode, its cluster, and “source” = first 2 chars
cell_df <- data.frame(
  cell_id     = cell_id,
  cluster_num = target,
  source      = substr(cell_id, 1, 2),  # take first 2 characters
  stringsAsFactors = FALSE
)

# 1c) Extract the raw counts matrix (genes × all cells)
raw_counts <- seruat_obj_22q@assays$RNA$counts
dim(raw_counts)  
# [1]    55 34765   # (55 genes, 34,765 cells)

# 1d) Subset the raw counts to only the cells in your target cluster
sub_counts <- raw_counts[, cell_id]

# e.g., [1]  55  1234  if target cluster has 1,234 cells



library(Matrix)

# 2a) Identify unique sources
sources <- unique(cell_df$source)
# e.g., sources might be c("C1", "C2", "C3", ..., "C9")

# 2b) Initialize a list to hold average vectors for each source
avg_list <- vector("list", length(sources))
names(avg_list) <- sources

# 2c) For each source, compute rowMeans of sub_counts restricted to cells of that source
for (src in sources) {
  #  1) Get the cell barcodes for this source
  cells_in_src <- cell_df$cell_id[cell_df$source == src]
  
  #  2) Subset sub_counts to those columns and compute the mean across each row (gene)
  #     Use rowMeans() on a sparse matrix → numeric vector length = nrow(sub_counts)
  #     (If sub_counts is large, Matrix::rowMeans() is efficient)
  avg_list[[src]] <- rowMeans(sub_counts[, cells_in_src, drop = FALSE])
}

# 2d) Combine the list of named numeric vectors into a single matrix
avg_by_source <- do.call(cbind, avg_list)
# Now dim(avg_by_source) = (nrow(sub_counts) genes) × (length(sources)).

# 2e) Assign rownames and colnames
rownames(avg_by_source) <- rownames(sub_counts)
colnames(avg_by_source) <- sources

# 2f) Inspect the result
# dim(avg_by_source)


avg_counts_mat <-avg_by_source 
# filterout the rows with less than 0.1 in all columns
avg_counts_mat <- avg_counts_mat[rowMeans(avg_counts_mat > 0) > 0, ]

# filter out the rowname not in FAM230F and SLC7A4
avg_counts_mat <- avg_counts_mat[! (rownames(avg_counts_mat) %in% c("FAM230F", "SLC7A4","P2RX6")), ]

# change the column name 
source_correct <- c("006", "008", "22q002", "22q005")
  
  
colnames(avg_counts_mat) <- source_correct
  
  
  
  
  
  
  gene_list <- rownames(avg_counts_mat)
  figure_folder <- sprintf("result/03-Figure-06-03/22q_genes/")
  
  plot_gene_heatmap(
  counts_mat      = avg_counts_mat,
  annotation_col  = annotation_col,
  gene_list       = gene_list,
  figure_folder   = figure_folder,
  file_name       = sprintf("avg_counts_heatmap_%s_row", target),
  reference_group = "CTRL",
  compare_group   = "22q11DS",
  cluster_rows    = FALSE,
  cluster_cols    = FALSE,
  save            = TRUE,
  fig.width       = 6,
  fig.height      = 10,
  scale           = "row",
  show_rownames   = TRUE)
  
  print("")
  
    plot_gene_heatmap(
  counts_mat      = avg_counts_mat,
  annotation_col  = annotation_col,
  gene_list       = gene_list,
  figure_folder   = figure_folder,
  file_name       = sprintf("avg_counts_heatmap_%s", target),
  reference_group = "CTRL",
  compare_group   = "22q11DS",
  cluster_rows    = FALSE,
  cluster_cols    = FALSE,
  save            = TRUE,
  fig.width       = 6,
  fig.height      = 10,
  scale           = "none",
  show_rownames   = TRUE)
  
  


  
}

```

\newpage
# 6. Pathway analysis
```{r pathway_analysis, fig.width=10, fig.height=6}
# get the file list from the data/synaptosomes_scRNA/cluster_DEG/WN_pathways_0603 folder
file_list <- list.files(here::here("data", "synaptosomes_scRNA", "cluster_DEG", "WN_pathways_0603"), full.names = TRUE)

# Load required libraries
library(ggplot2)
library(dplyr)

# Assume file_list is already defined, e.g.:
# file_list <- c(
#   "/…/Deep_layer_excitatory_neurons_down_enrichment_WN.csv",
#   "/…/Deep_layer_excitatory_neurons_up_enrichment_WN.csv",
#   "/…/Excitatory_neurons_down_enrichment_WN.csv",
#   "/…/Excitatory_neurons_up_enrichment_WN.csv",
#   "/…/Immature_preoptic_area_(pOA)_interneurons_down_enrichment_WN.csv",
#   "/…/Immature_preoptic_area_(pOA)_interneurons_up_enrichment_WN.csv",
#   "/…/Interneuron_down_enrichment_WN.csv",
#   "/…/Interneuron_up_enrichment_WN.csv",
#   "/…/Upper_layer_excitatory_neurons_down_enrichment_WN.csv",
#   "/…/Upper_layer_excitatory_neurons_up_enrichment_WN.csv"
# )

# Define the folder where you want to save the plots
result_folder <- "result/03-Figure-06-03/cluster_DEG_enrichment/"  # <-- change this to your desired output directory

# Loop over each enrichment‐CSV in file_list
for (file_path in file_list) {
  # 1. Read in the CSV
  df <- read.csv(file_path, stringsAsFactors = FALSE)
  
  # 2. Take the top 30 rows by ascending p_value (assuming smaller p_value = more significant)
  result_df_30 <- df %>%
    arrange(p_value) 
  # drop the NA rows
  result_df_30 <- result_df_30[!is.na(result_df_30$p_value), ]
    result_df_30 <- result_df_30 %>%
    mutate(term_name = factor(term_name, levels = rev(term_name)))
    
  # make the term_name as factor
  
  # 3. Determine "flag" (Up / Down) from the filename
  #    We assume filenames contain "_up_" or "_down_" (case‐insensitive).
  if (grepl("_up_", file_path, ignore.case = TRUE)) {
    flag <- "Up"
  } else if (grepl("_down_", file_path, ignore.case = TRUE)) {
    flag <- "Down"
  } else {
    flag <- "Other"
  }
  
  # 4. Extract a base name for the output files (without extension)
  #    Example: "Deep_layer_excitatory_neurons_down_enrichment_WN"
  file_name <- tools::file_path_sans_ext(basename(file_path))
  
  # 5. Build the ggplot bar chart
  p <- ggplot(result_df_30, aes(x = term_name,
                                y = -log10(p_value),
                                fill = -log10(p_value))) +
    geom_bar(stat = "identity", show.legend = TRUE) +
    # Apply different color scales based on flag value
    {
      if (flag == "Up") {
        scale_fill_gradient(low = "#fc9272", high = "#a50f15")  # Red scale for UP
      } else if (flag == "Down") {
        scale_fill_gradient(low = "#56B1F7", high = "#132B43")  # Blue scale for DOWN
      } else {
        scale_fill_manual(values = rep("lightgrey", nrow(result_df_30)))  # Default light grey
      }
    } +
    ylab("-log10(FDR)") +
    xlab("") +
    # ggtitle(paste0(file_name, "")) +
    coord_flip() +
    theme_classic(base_size = 16) +
    theme(
      text = element_text(family = "Arial", size = 16, colour = "black"),
      axis.title = element_text(color = "black"),
      axis.text = element_text(color = "black"),
      plot.title = element_text(family = "Arial", size = 16, face = "bold", hjust = 0.5)
    ) +
    geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "black")+
  theme(
    panel.background  = element_rect(fill = "transparent", colour = NA),
    plot.background   = element_rect(fill = "transparent", colour = NA),
    legend.background = element_rect(fill = "transparent", colour = NA)
  )
  
  # 6. Print the plot to the graphics device (optional)
  print(p)
  
  # 7. Save both PNG and PDF versions to result_folder
  ggsave(
    filename = file.path(result_folder, paste0(file_name, "_pathways.png")),
    plot     = p,
    width    = 10,
    height   = 6,
    units    = "in",
    dpi      = 300,
    bg       = "transparent"   # <-- this makes the PNG background transparent
  )
  ggsave(
    filename = file.path(result_folder, paste0(file_name, "_pathways.pdf")),
    plot     = p,
    width    = 10,
    height   = 6,
    units    = "in",
    bg       = "transparent" 
  )
}

```



\newpage
Session Information
```{r session_info}
sessionInfo()
```
